Jsontoolpath Generation for Makerbot 5th Gen
Added JSONToolpath generation for Makerbot 5th Gen printers, it currently outputs only the toolpath, not the metadata.
Will need to pass the extruder temp settings + other metadatas to have CuraEngine do all the work of outputting the files, so that nothings happens in Cura.
Currently, Cura adds the extruder temperature in python (checks if start.gcode contains the {print_temperature} flag, and adds a M109 if not), this needs to be fixed.