Passing temperature settings to CuraEngine config
Passing the temperatures from the config file from Cura to CuraEngine allows us to write the extruder temp + metadata for the Makerbot 5th Gen printers.
json toolpath + metadata are passed to Cura as a single file, and Cura cuts it in the middle to save both files + add the default pictures from ressource folder & zip everything.